renamed azureAuth
