+++ src/extension/src/azure-auth/azureAuth.ts
