+++ src/extension/src/reactPanel.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import * as path from "path";
-import { AzureAuth } from './azure-auth/AzureAuth';
+import { AzureAuth } from './azure-auth/azureAuth';
 
 /**
  * Manages react webview panels
